moved into /lesson and added LandingPage
