+++ src/app/econ/supply-demand/lesson/SupplyDemandLesson.tsx
